Added some small tests for login on client. LogIn page for login and small navbar change on index.

+++ Oblig-1-ITPE3200/Controllers/ObligController.cs
@@ -1,7 +1,9 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
 using Oblig_1_ITPE3200.DAL;
 using Oblig_1_ITPE3200.DTOs;
 using Oblig_1_ITPE3200.Models;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -11,6 +13,8 @@ namespace Oblig_1_ITPE3200.Controllers
     public class ObligController : ControllerBase
     {
         private readonly IObligRepository _db;
+
+        private const string _loggedOn = "loggedOn";
 
         public ObligController(IObligRepository db)
         {
@@ -24,19 +28,39 @@ namespace Oblig_1_ITPE3200.Controllers
         }
         public async Task<List<DiseaseDTO>> GetAllDiseases(string searchString)
         {
-            return await _db.GetAllDiseases(searchString);
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedOn)))
+            {
+                return Unauthorized();
+            }
+
+            List<DiseaseDTO> diseaseDTO = await _db.GetAllDiseases(searchString);
+            
         }
         public async Task<bool> CreateDisease(Disease disease)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedOn)))
+            {
+                return Unauthorized();
+            }
+
             return await _db.CreateDisease(disease);
         }
 
         public async Task<bool> UpdateDisease(Disease disease)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedOn)))
+            {
+                return Unauthorized();
+            }
+
             return await _db.UpdateDisease(disease);
         }
         public async Task<bool> DeleteDisease(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedOn)))
+            {
+                return Unauthorized();
+            }
             return await _db.DeleteDisease(id);
         }
 
@@ -87,6 +111,63 @@ namespace Oblig_1_ITPE3200.Controllers
         public async Task<List<DiseaseDTO>> SearchDiseases(int[] symptomIds)
         {
             return await _db.SearchDiseases(symptomIds);
+        }
+
+        public async Task<ActionResult> LogIn(User user)
+        {
+            try
+            {
+                if (ModelState.IsValid)
+                {
+                    bool b = await _db.LogIn(user);
+                    if (!b)
+                    {
+                        //_log.LogInformation("Log in failed with user=" + user.Username);
+                        HttpContext.Session.SetString(_loggedOn, "");
+                        return Ok(false);
+                    }
+                    HttpContext.Session.SetString(_loggedOn, "LoggedOn");
+                    return Ok(true);
+                }
+                //_log.LogInformation("Something wrong in inputvalidation, user=" + user.Username);
+                return BadRequest("Something wrong in inputvalidation");
+            }
+            catch (Exception e)
+            {
+                //_log.LogInformation(e.Message);
+                return BadRequest("Something wrong on server.");
+            }
+        }
+
+        public bool LogOut()
+        {
+            try
+            {
+                HttpContext.Session.SetString(_loggedOn, "");
+                return true;
+            }
+            catch (Exception e)
+            {
+                //_log.LogInformation(e.Message);
+                return false;
+            }
+        }
+
+        public ActionResult IsLoggedIn()
+        {
+            try
+            {
+                if (HttpContext.Session.GetString(_loggedOn) == "LoggedOn")
+                {
+                    return Ok(true);
+                }
+                return Ok(false);
+            }
+            catch (Exception e)
+            {
+                //_log.LogInformation(e.Message);
+                return BadRequest("Something wrong with server. Try again later.");
+            }
         }
     }
 }
